Decimal number formatting

+++ Modulo 6/Lista 01/SimpleCalculator/src/com/letscode/service/Calculator.java
@@ -13,7 +13,7 @@ public class Calculator {
     return numbers[0] * numbers[1];
   }
 
-  public static int divide(int[] numbers) {
-    return numbers[0] / numbers[1];
+  public static float divide(int[] numbers) {
+    return (float) numbers[0] / numbers[1];
   }
 }
